+++ NEWS.md
@@ -1,5 +1,6 @@
 # pals 1.7 - Unpublished
 
+* Added 'turbo' palette.
 
 # pals 1.6 - Dec 2019